+++ src/models/httpFile.ts
@@ -8,7 +8,6 @@ export interface HttpFile {
   readonly hooks: HttpFileHooks;
   readonly httpRegions: Array<HttpRegion>;
   readonly globalHttpRegions: Array<HttpRegion>;
-  activeEnvironment?: string[];
 
   findHttpRegion(name: string): HttpRegion | undefined;
 }