+++ src/test/assert/javascript.spec.ts
@@ -17,8 +17,11 @@ describe('assert.javascript', () => {
     ?? js response.parsedBody == {"foo": "bar"}
     `);
 
-    const responses = await sendHttpFile(httpFile, {
-      host: `http://localhost:${localServer.port}`,
+    const responses = await sendHttpFile({
+      httpFile,
+      variables: {
+        host: `http://localhost:${localServer.port}`,
+      },
     });
     expect(responses.length).toBe(1);
     expect(responses[0].statusCode).toBe(200);